refactor(sdk): move builders to internal and rename
- Move builders from testing/ to internal/builders.ts
- Rename MockPrivateTransfersBuilder → PrivateTransfersBuilderImpl
- Rename MockTokenOperationsBuilder → TokenOperationsBuilderImpl
- Change builder execute() return type from CallAndProof[] to CallAndProof
- Separate builder tests to tests/internal/builders.test.ts
- Update testing/transfers.ts to use new builder from internal
