+++ sdk/src/testing/index.ts
@@ -5,7 +5,6 @@
 export { ERC20, ERC20s } from "./erc20.js";
 export { PrivacyPool } from "./pool.js";
 export { MockDiscoveryProvider } from "./discovery.js";
-export { MockTokenOperationsBuilder, MockPrivateTransfersBuilder } from "./builders.js";
 export { MockPrivateTransfers } from "./transfers.js";
 export { createMockProof, createMockCallAndProof, Withdrawal } from "./helpers.js";
 export { hashes } from "../utils/hashes.js";
